requirements.txt
fastapi
uvicorn
python-multipart
torch
numpy
opencv-python-headless
albumentations
timm
Pillow
jinja2